requirements.txt
opencv-python>=4.5.2.54
torchvision>=0.9.1
Pillow>=8.2.0
numpy>=1.19.5
torch>=1.8.1
tqdm>=4.61.0
scipy>=1.6.3
matplotlib>=3.2.2
prettytable>=2.1.0
thop>=0.0.31.post2005241907
setuptools>=57.0.0
tensorboardX>=2.2
lpips>=0.1.3